Add linear and random agent
Initial versions of the linear and random agents created. Still need to
setup passing in arguments. Currently the linear agent uses the
Dellacherie weights as its default, its not quick enough to finish one
of these games in good time. Currently can clear about 60 lines a second
which is 10,000X slower than the linear mdptetris implementation in C.
Not sure how this could be sped up to that level!