Create new structure and add linear function
Created a new structure for the module to allow reuseability and pip
install setup. Need to create requisite `setup.py` file in root
directory.

Added doc strings to the LinearGame file, and also a `get_next_states`
function useful for state-value function approximations. Potentially
should add the functionality as an environment in gym_mdptetris, but
it's good to have here. This allows an experiment or an agent to
initialise a LinearGame and get the subsequent states as a dictionary.
The agent can then use a method to approximately compute the best
subsequent state-action value, and select the action that results in
that state. This should be universally applicable whether a heuristic
representation of a state is given, or whether the underlying state
array is used. The experiments will then seek to discover if learning is
possible when using the underlying arrays as well as heuristic versions.