Add saving
Add compatability for saving, including adding a savedir, and a runid
(the current ISO8601 date-time, may have to vary that).
A better runid (if runs are getting started at the same second) would be
a hash of the args? Otherwise I'll just have to make sure I kick off any
runs at least a second apart. There's probably a better way, but this is
at least one way.