Initial 1D DQN implementation
Created the initial implementation of DQN for the 1D state space
version. This is where the agent approximates the state value function,
taking as input the board array flattened into 1D. This was showing
successful learning along the lines of the dellacherie state
approximation!

The implementation of the alternative state spaces is pretty ugly, so
that can definitely use cleaning up. As a brief hack I put in a
dictionary of the setup functions, so passing in "1D" for the state
representation pulls the required neural net architecture and
environment in. It's hacky but functional.