Fake multiprocessing for PPO initial
Adding a fake multiprocessing class for PPO. This will create a number
of Tetris games, create stacks of states to pass through the GPU, and
stack states for the agent. Just created the skeleton out of functions
that are used in the DQN agent, need to make the initial work on the
skeleton of PPO to decide if anything else is required, and how this
class will work.