Update DQN with input dims, allow diff board size
Updated DQN with input dimension options to enable differing board
sizes with experiments. Previously the NN was not compatible with this.
It checks if the state representation is heuristic, in which case it
sets the input dimension to 6, otherwise it passes in the board height
multiplied by the board width.