Chnage NN file
Move NN file to be generic instead of DQN specific, and update the
training as a result. This will be useful for PPO using the same
networks.

Also modified the NN to use the recommended implementation of feed
forward instead of having the relu in place.

Other change made in this is to add a few more docs to the arg parser
and add a save function.