Simplify 1D game
Simplified the methods of the 1D game. Since the only method that is
actually different is the `get_state` method, removed other functions
for greater simplicity.

Thinking about whether the DQN agent could be rewritten to function for
state-action-values as well as just state-values. Need to decide whether
to pursue state-action-values, or if it is possible to write PPO to
function on just state-values with its actor/critic distinction.