PPO update
Add functions to PPO, including skeleton for get action, and start on
rollouts. Also adding initialisation of hyper parameters. Not sure where
to set defaults, either in an initialisation func or in the argparser to
get the values. Moved argparser to a main file where PPO can be called
from.

Still thinking about modularising DQN, and applying DQN to a model-free
environment.