Fix bug in epsilon decay
When I added in the linear epsilon decay I set the epsilon check to be
`min` instead of `max` on verifying the epsilon was above the final.
Seems like it might be learning even when skipping to the final epsilon?
Will leave this run (id: 20210802T203842Z) to see what happens, and then
restart with correction.