Saving and loading for PPO
Just redid saving and loading for PPO, need to test that the loading
actually functions. This is a breaking hcange with previously trained
models, so I may need to re-run my experiments to get fresh models.
Luckily that only takes 24hrs at the moment. Need to do the same for DQN
which may involve rerunning all of these too...