TD-Lambda introduction
The strategy I have previously been referring to as DQN is actually a
modified version of TD-Lambda. I've now created a directory for this and
renamed appropriately. It is quite successful on the testing! I've
ironed out a bug in the test strategy and am now running it on an old
trained model, but with zero epsilon it is quite successful.