Sort smoothing
Found functions that enable graph smoothing. First there's the option to
group the data up in the dataframe, so the average over 100 steps or so.
Second you can plot with a smoothed alpha, and then plot the actual
values in a lighter colour. I need to parameterise these so they're
reuseable.